+++ src/components/index.js
@@ -1,4 +1,3 @@
 export * from "./Player";
 export { default as HostPlayer } from "./HostPlayer";
 export { default as ListenerPlayer } from "./ListenerPlayer";
-export * from "./Donations";
